Move function to where it's used
